Add notebooks showing how to set up STAC objects (#63)
