+++ _quarto.yml
@@ -38,34 +38,34 @@ website:
           - services/dashboard.qmd
           - services/apis.qmd
           - services/jupyterhub.qmd
-      - section: example-notebooks/index.qmd
+      - section: notebooks/index.qmd
         contents:
           - section: Quickstarts
             contents:
               - section: Accessing the Data Directly
                 contents:
-                  - example-notebooks/list-collections.ipynb
-                  - example-notebooks/open-and-plot.ipynb
-                  - example-notebooks/timeseries-rioxarray-stackstac.ipynb
-                  - example-notebooks/visualize-zarr.ipynb
-                  - example-notebooks/visualize-multiple-times.ipynb
-                  - example-notebooks/downsample-zarr.ipynb
-                  - example-notebooks/intake.ipynb
-                  - example-notebooks/wfs.ipynb
+                  - notebooks/quickstarts/list-collections.ipynb
+                  - notebooks/quickstarts/open-and-plot.ipynb
+                  - notebooks/quickstarts/timeseries-rioxarray-stackstac.ipynb
+                  - notebooks/quickstarts/visualize-zarr.ipynb
+                  - notebooks/quickstarts/visualize-multiple-times.ipynb
+                  - notebooks/quickstarts/downsample-zarr.ipynb
+                  - notebooks/quickstarts/intake.ipynb
+                  - notebooks/quickstarts/wfs.ipynb
               - section: Using the Raster API
                 contents:
-                  - example-notebooks/no2-map-plot.ipynb
-                  - example-notebooks/timeseries-stac-api.ipynb
-                  - example-notebooks/hls-visualization.ipynb
+                  - notebooks/quickstarts/no2-map-plot.ipynb
+                  - notebooks/quickstarts/timeseries-stac-api.ipynb
+                  - notebooks/quickstarts/hls-visualization.ipynb
           - section: Tutorials
             contents:
-              - example-notebooks/gif-generation.ipynb
+              - notebooks/tutorials/gif-generation.ipynb
           - section: Datasets
             contents:
-              - example-notebooks/ocean-npp-timeseries-analysis.ipynb
-              - example-notebooks/nceo-biomass-statistics.ipynb
-              - example-notebooks/volcano-so2-monitoring.ipynb
-              - example-notebooks/air-quality-covid.ipynb
+              - notebooks/datasets/ocean-npp-timeseries-analysis.ipynb
+              - notebooks/datasets/nceo-biomass-statistics.ipynb
+              - notebooks/datasets/volcano-so2-monitoring.ipynb
+              - notebooks/datasets/air-quality-covid.ipynb
       - section: contributing/index.qmd
         contents:
           - contributing/docs-and-notebooks.qmd
@@ -73,8 +73,14 @@ website:
             contents:
               - contributing/dataset-ingestion/file-preparation.qmd
               - contributing/dataset-ingestion/catalog-ingestion.qmd
-              - contributing/dataset-ingestion/stac-collection-conventions.qmd
-              - contributing/dataset-ingestion/stac-item-conventions.qmd
+              - section: STAC Collection Creation
+                contents:
+                - contributing/dataset-ingestion/stac-collection-conventions.qmd
+                - example_notebooks/veda_operations/stac-collection-creation.ipynb
+              - section: STAC Item Creation
+                contents:
+                - contributing/dataset-ingestion/stac-item-conventions.qmd
+                - example_notebooks/veda_operations/stac-item-creation.ipynb
           - section: contributing/dashboard-configuration/index.qmd
             contents:
               - contributing/dashboard-configuration/dataset-configuration.qmd
