--- example-notebooks/index.qmd
@@ -1,42 +1,0 @@
----
-title: Usage Examples
----
-
-## Getting started
-The example notebooks are divided into three sections: 
-
-- **Quickstarts**: Notebooks to get you started quickly and help you become more familiar with cloud-native geospatial technologies.
-- **Tutorials**: Longer notebooks that walk through more advanced use cases and examples.
-- **Datasets**: Notebooks that showcase a particular VEDA dataset and walk through an applied geospatial analyses.
-
-### Choosing the right data access route for your needs
-The **Quickstarts** examples are further divided into two sections, which you can choose from depending on your data needs:
-
-- **Accessing the Data Directly**: For when you want to access the raw data (e.g., to do a specific analysis). In this case, permissions are required to access the data (i.e., must be run on VEDA JupyterHub) and computation happens within the user's instance (i.e., the user needs to think about instance size). This approach is suitable for use within notebooks. All examples provided in this section require VEDA JupyterHub access to run. 
-
-- **Using the Raster API**: For when you want to show outputs to other people or do standard processing. No permissions required (i.e., notebooks can be run on `mybinder`). Additionally, the computation happens somehwere else (i.e., user does not have to think about instance size). Lastly, this approach is suitable for use within notebooks as well as web application frontends (e.g., like dataset discoveries). These notebook examples can be run on both VEDA JupyterHub, as well as outside of the Hub (see instructions below) and within `mybinder`.
-
-
-## How to run
-
-Every notebook contains information about how to run it. Some can run on [mybinder](https://mybinder.org/) and all can run on the VEDA JupyterHub. See [VEDA Analytics JupyterHub Access](https://nasa-impact.github.io/veda-docs/veda-jh-access.html) for information about how to gain access.
-
-### Running outside of VEDA JupyterHub
-
-To run the notebooks locally, you can use can install the Python packages (a [virtual environment](https://docs.python.org/3/library/venv.html) is recommended)
-
-```bash
-pip install -r requirements.txt
-```
-
-Once you have installed the packages you can run the notebooks using [Jupyter](https://jupyter.org/).
-
-```bash
-jupyter lab
-```
-
-If the notebook needs access to protected data on S3, you will need to specifically get access. Please request access by emailing aimee@developmentseed.org or alexandra@developmentseed.org and providing your affiliation, interest in or expected use of the dataset and an AWS IAM role or user Amazon Resource Name (ARN).
-
-## How to contribute
-
-Please refer to the [notebook style guide in these docs](../contributing/docs-and-notebooks.qmd).
