+++ services/jupyterhub.qmd
@@ -22,6 +22,6 @@ Access to the VEDA notebook environment is currently on an as-need basis. If you
 * Once you accepted the invitation, you should be able to go to [https://nasa-veda.2i2c.cloud/](https://nasa-veda.2i2c.cloud/) and login via your Github credentials.
 
 
-## Example notebooks
+## Instructory notebooks
 
-This documentation site provides [example notebooks](../example-notebooks/index.qmd) on how to load and analyze Earth data an interactive cloud computing environment.
+This documentation site provides [Jupyter notebooks](../notebooks/index.qmd) on how to load and analyze Earth data an interactive cloud computing environment.
